+++ noodles-cram/src/record/convert.rs
@@ -34,8 +34,7 @@ impl Record {
         }
 
         if let Some(name) = record.name() {
-            let name = sam::alignment::record_buf::Name::from(name.as_bytes());
-            builder = builder.set_name(name);
+            builder = builder.set_name(name.into());
         }
 
         // next mate bit flags